% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\docType{methods}
\name{outputvars}
\alias{outputvars}
\alias{outputvars,ProvStoreDF-method}
\alias{outputvars,ProvStoreDF}
\alias{outputvarhashes}
\alias{outputvarhashes,ProvStoreDF-method}
\alias{outputvarhashes,ProvStoreDF}
\alias{outputvarclasses}
\alias{outputvarclasses,ProvStoreDF-method}
\alias{outputvarclasses,ProvStoreDF}
\alias{inputvars}
\alias{inputvars,ProvStoreDF-method}
\alias{inputvars,ProvStoreDF}
\alias{inputvarhashes}
\alias{inputvarhashes,ProvStoreDF-method}
\alias{inputvarhashes}
\alias{inputvarclasses}
\alias{inputvarclasses,ProvStoreDF-method}
\alias{inputvarclasses,ProvStoreDF}
\alias{provdata}
\alias{provdata,ProvStoreDF-method}
\alias{provdata,ProvStoreDF}
\alias{hashprefix}
\alias{hashprefix,ProvStoreDF-method}
\alias{hashprefix,ProvStoreDF}
\title{Accessors}
\usage{
outputvars(obj)

\S4method{outputvars}{ProvStoreDF}(obj)

outputvarhashes(obj)

\S4method{outputvarhashes}{ProvStoreDF}(obj)

outputvarclasses(obj)

\S4method{outputvarclasses}{ProvStoreDF}(obj)

inputvars(obj)

\S4method{inputvars}{ProvStoreDF}(obj)

inputvarhashes(obj)

\S4method{inputvarhashes}{ProvStoreDF}(obj)

inputvarclasses(obj)

\S4method{inputvarclasses}{ProvStoreDF}(obj)

provdata(obj)

\S4method{provdata}{ProvStoreDF}(obj)

hashprefix(obj)

\S4method{hashprefix}{ProvStoreDF}(obj)
}
\arguments{
\item{obj}{The object.}
}
\description{
Accessors for information in roprov objects
}
\examples{
library(fastdigest)
code = c("x = 5", "y = x + 1")
outvars = c("x", "y")
outvarhashes = c(fastdigest(5), fastdigest(6))
outvarclasses = rep("numeric", 2)
invars = c("", "x")
invarhashes = c("", fastdigest(5))
invarclasses = c("", "numeric")

ps = ProvStoreDF(outvars = outvars,
                 outvarhashes = outvarhashes,
                 outvarclasses = outvarclasses,
                 invars = invars,
                 invarhashes = invarhashes,
                 invarclasses = invarclasses,
                 code = code,
                 agent = "coolguyorgirl")
outputvars(ps)
outputvarhashes(ps)
outputvarclasses(ps)
inputvars(ps)
inputvarhashes(ps)
inputvarclasses(ps)
provdata(ps)
hashprefix(ps)
}
