% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{cback1}
\alias{cback1}
\title{alloutputs}
\usage{
cback1(invar, prevouts, prevouthashes, prevoutclasses)
}
\value{
a character vector of all variables which were created or modified
by the expression corresponding to \code{x}
}
\description{
Get all output variables (outputs + updates) from a ScriptNodeInfo object
}
