% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{ProvStores}
\alias{ProvStores}
\title{Combine provenance stores}
\usage{
ProvStores(...)
}
\arguments{
\item{\dots}{Two or more ProvStoreDF objects}
}
\value{
A ProvStoreDF object
}
\description{
This operation is conceptually
}
