% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\docType{methods}
\name{rbind,ProvStoreDF-method}
\alias{rbind,ProvStoreDF-method}
\title{rbind method}
\usage{
\S4method{rbind}{ProvStoreDF}(..., deparse.level = 1)
}
\arguments{
\item{\dots}{Two or more ProvStoreDF objects. Must all have
identical hashprefix values}

\item{deparse.level}{ignored.}
}
\value{
A ProvStoreDF object.
}
\description{
an rbind method for ProvStoreDF objects
}
