% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{outputvars}
\alias{outputvars}
\alias{outputvarhashes}
\alias{outputvarclasses}
\alias{inputvars}
\alias{inputvarhashes}
\alias{inputvarclasses}
\alias{provdata}
\alias{hashprefix}
\title{Accessors}
\usage{
outputvars(obj)

outputvarhashes(obj)

outputvarclasses(obj)

inputvars(obj)

inputvarhashes(obj)

inputvarclasses(obj)

provdata(obj)

hashprefix(obj)
}
\arguments{
\item{obj}{The object.}
}
\description{
Accessors for information in roprov objects.
}
