% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{cback1}
\alias{cback1}
\title{cback1}
\usage{
cback1(invar, prevouts, prevouthashes, prevoutclasses)
}
\arguments{
\item{invar}{The variable to find}

\item{prevouts}{Previous output variables}

\item{prevouthashes}{Hashes for the previous output variables}

\item{prevoutclasses}{Classes for the previous output variables}
}
\value{
a list containing the hash and class of the variable.
}
\description{
Determine the hash and class of a varaible based on
    the most recent ouput values for that variable captured as a
    previous output. Not intended for direct use by end users.
}
\details{
If the variable is not found in the previous outputs, the
    .Global environment is inspected for a current value of the
    variable. This is necessary to capture things that can show up
    via lazyloading without ever being an output v, e.g., the
    mtcars dataset.
}
